Priors for `rsstap` models

Priors for `rsstap` models

GLMs

Using one spatial aggregated predictor as an example, stap_(g)lm models have the following form.

$$ g(\mu_i) = Z_i^T \delta + \sum_d \sum_l \beta_l\phi_l(d) $$

Currently the priors in rsstap are fixed and are always of the following form:

$$p(\delta) \propto 1$$

$$\sigma \sim C^+(0,5)$$

$$\beta \sim MVN_L(0,\sum_k S_k \tau_k)$$

$$\tau_k \sim \text{Exponential}(1)$$

Where \(S_k\) are generated from the jagam function and sum to form a complete precision matrix with different \(\tau\) penalties along the diagonal.

GLMERs

Using only one spatial aggregated predictor as an example, stap_(g)lmer models have the following form:

$$ g(\mu_{ij}) = Z_{ij}^T \delta + \sum_d \sum_l \beta_l\phi_l(d) + W_{ij}^Tb_i $$

Where

$$b_i \sim N(0,\Sigma)$$

priors for \(\delta\),\(\beta\),\(\sigma\),\(\tau_k\) are the same as before, but now \(\Sigma\) is decomposed as described here.