Methods for sstapreg objects

The methods documented on this page are actually some of the least important methods defined for sstapreg objects. The most important methods are documented separately, each with its own page. Links to those pages are provided in the See Also section, below.

# S3 method for sstapreg
coef(object, ...)

# S3 method for sstapreg
confint(object, ...)

# S3 method for sstapreg
fitted(object, ...)

# S3 method for sstapreg
se(object, ...)

# S3 method for sstapreg
nobs(object, ...)

# S3 method for sstapreg
VarCorr(x)

# S3 method for sstapreg
ngrps(object, ...)

# S3 method for sstapreg
vcov(object, correlation = FALSE, ...)

# S3 method for sstapreg
nsamples(object, ...)

Arguments

object
sstapreg object

...
Ignored

x
sstapreg object

correlation
For vcov, if FALSE (the default) the covariance matrix is returned. If TRUE, the correlation matrix is returned instead.

Details

The methods documented on this page are similar to the methods defined for objects of class 'lm', 'glm', 'glmer', etc. However there are a few key differences:

coef

Medians are used for fixed effect point estimates. See the Point estimates section in print.sstapreg for more details.

se

The se function returns fixed effects standard errors based on mad. See the Uncertainty estimates section in print.sstapreg for more details.

confint

confint will throw an error because the posterior_interval function should be used to compute Bayesian uncertainty intervals.